feat: Storage refactoring
- Move prefix wrapper folder from pkg/storage to pkg/store. The pkg/storage folder is now gone.
- Added clarification in Iterator Next() comments regarding the expected behaviour when the iterator is exhausted.
- Minor refactor in mobile storage wrapper.
